+++ test/unit/handlers/v1/logging_test.py
@@ -1,15 +1,17 @@
-import unittest
+# -*- coding: utf-8 -*-
 import json
+import unittest
 
-from . import TestHandlerBase
 from mock import patch
+
+from .. import TestHandlerBase
 
 
 class LoggingApiTest(TestHandlerBase):
     def setUp(self):
         super(LoggingApiTest, self).setUp()
 
-    @patch("brew_view.controllers.logging_api.MongoParser.serialize_logging_config")
+    @patch("brew_view.handlers.v1.logging.MongoParser.serialize_logging_config")
     def test_get_config(self, serialize_mock):
         serialize_mock.return_value = "serialized_logging_config"
 
@@ -18,7 +20,7 @@ class LoggingApiTest(TestHandlerBase):
         self.assertEqual("serialized_logging_config", response.body.decode("utf-8"))
 
     @patch("brew_view.load_plugin_logging_config")
-    @patch("brew_view.controllers.system_api.MongoParser.serialize_logging_config")
+    @patch("brew_view.handlers.v1.system.MongoParser.serialize_logging_config")
     def test_patch_reload(self, serialize_mock, load_mock):
         serialize_mock.return_value = "serialized_logging_config"
 
@@ -32,7 +34,7 @@ class LoggingApiTest(TestHandlerBase):
         self.assertEqual("serialized_logging_config", response.body.decode("utf-8"))
         self.assertEqual(load_mock.call_count, 1)
 
-    @patch("brew_view.controllers.system_api.MongoParser.serialize_logging_config")
+    @patch("brew_view.handlers.v1.system.MongoParser.serialize_logging_config")
     def test_patch_invalid_operation(self, serialize_mock):
         body = json.dumps({"operations": [{"operation": "INVALID"}]})
         serialize_mock.return_value = "serialized_logging_config"
