+++ test/unit/handlers/v1/request_test.py
@@ -1,5 +1,8 @@
+# -*- coding: utf-8 -*-
+import copy
 import json
 
+import datetime
 import pytest
 from mock import MagicMock, Mock, PropertyMock, patch
 from tornado.gen import Future
@@ -7,11 +10,12 @@ from tornado.httpclient import HTTPRequest, HTTPClientError
 
 import bg_utils
 import brew_view
-from bg_utils.mongo.models import Request, System
-from brew_view.controllers import RequestListAPI
+from bg_utils.mongo.models import Request, Job, RequestTemplate, DateTrigger
+from bg_utils.mongo.models import System
+from brew_view.handlers.v1.request import RequestListAPI
 from brewtils.rest.client import RestClient
 from brewtils.schema_parser import SchemaParser
-from . import TestHandlerBase
+from .. import TestHandlerBase
 
 
 @pytest.fixture(autouse=True)
@@ -29,7 +33,7 @@ def clear_requests(app):
 def thrift(monkeypatch, thrift_context, thrift_client, process_future):
     thrift_client.processRequest.return_value = process_future
     monkeypatch.setattr(
-        brew_view.controllers.request_list_api, "thrift_context", thrift_context
+        brew_view.handlers.v1.request, "thrift_context", thrift_context
     )
 
 
@@ -37,7 +41,7 @@ def thrift(monkeypatch, thrift_context, thrift_client, process_future):
 def latency_total(monkeypatch):
     latency_total = Mock()
     monkeypatch.setattr(
-        brew_view.controllers.request_list_api, "http_api_latency_total", latency_total
+        brew_view.handlers.v1.request, "http_api_latency_total", latency_total
     )
     return latency_total
 
@@ -65,6 +69,311 @@ def post_bg_request(base_url, bg_request):
         headers=RestClient.JSON_HEADERS,
         body=SchemaParser.serialize_request(bg_request, to_string=True),
     )
+
+
+class RequestAPITest(TestHandlerBase):
+    def setUp(self):
+        self.request_mock = Mock()
+
+        self.ts_epoch = 1451606400000
+        self.ts_dt = datetime.datetime(2016, 1, 1)
+        self.request_dict = {
+            "children": [],
+            "parent": None,
+            "system": "system_name",
+            "system_version": "0.0.1",
+            "instance_name": "default",
+            "command": "say",
+            "id": "58542eb571afd47ead90d25f",
+            "parameters": {},
+            "comment": "bye!",
+            "output": "nested output",
+            "output_type": "STRING",
+            "status": "IN_PROGRESS",
+            "command_type": "ACTION",
+            "created_at": self.ts_epoch,
+            "updated_at": self.ts_epoch,
+            "error_class": None,
+            "metadata": {},
+            "has_parent": True,
+            "requester": None,
+        }
+        self.job_dict = {
+            "name": "job_name",
+            "trigger_type": "date",
+            "trigger": {"run_date": self.ts_epoch, "timezone": "utc"},
+            "request_template": {
+                "system": "system",
+                "system_version": "1.0.0",
+                "instance_name": "default",
+                "command": "speak",
+                "parameters": {"message": "hey!"},
+                "comment": "hi!",
+                "metadata": {"request": "stuff"},
+            },
+            "misfire_grace_time": 3,
+            "coalesce": True,
+            "next_run_time": self.ts_epoch,
+            "success_count": 0,
+            "error_count": 0,
+        }
+        db_dict = copy.deepcopy(self.job_dict)
+        db_dict["request_template"] = RequestTemplate(**db_dict["request_template"])
+        db_dict["trigger"]["run_date"] = self.ts_dt
+        db_dict["trigger"] = DateTrigger(**db_dict["trigger"])
+        db_dict["next_run_time"] = self.ts_dt
+        self.job = Job(**db_dict)
+
+        db_dict = copy.deepcopy(self.request_dict)
+        db_dict["created_at"] = self.ts_dt
+        db_dict["updated_at"] = self.ts_dt
+        self.request = Request(**db_dict)
+
+        super(RequestAPITest, self).setUp()
+
+    def tearDown(self):
+        Request.objects.delete()
+        Job.objects.delete()
+
+    def test_get(self):
+        self.request.save()
+        response = self.fetch("/api/v1/requests/" + str(self.request.id))
+        self.assertEqual(200, response.code)
+        data = json.loads(response.body.decode("utf-8"))
+        data.pop("updated_at")
+        self.request_dict.pop("updated_at")
+        self.assertEqual(self.request_dict, data)
+
+    def test_patch_replace_duplicate(self):
+        self.request.status = "SUCCESS"
+        self.request.output = "output"
+        self.request.save()
+        body = json.dumps(
+            {
+                "operations": [
+                    {"operation": "replace", "path": "/output", "value": "output"},
+                    {"operation": "replace", "path": "/status", "value": "SUCCESS"},
+                ]
+            }
+        )
+
+        response = self.fetch(
+            "/api/v1/requests/" + str(self.request.id),
+            method="PATCH",
+            body=body,
+            headers={"content-type": "application/json"},
+        )
+        self.assertEqual(200, response.code)
+
+        self.request.reload()
+        self.assertEqual("SUCCESS", self.request.status)
+        self.assertEqual("output", self.request.output)
+
+    def test_patch_replace_status(self):
+        self.request.save()
+        body = json.dumps(
+            {
+                "operations": [
+                    {"operation": "replace", "path": "/status", "value": "SUCCESS"}
+                ]
+            }
+        )
+
+        response = self.fetch(
+            "/api/v1/requests/" + str(self.request.id),
+            method="PATCH",
+            body=body,
+            headers={"content-type": "application/json"},
+        )
+        self.assertEqual(200, response.code)
+        self.request.reload()
+        self.assertEqual("SUCCESS", self.request.status)
+
+    def test_patch_replace_output(self):
+        self.request.output = "old_output_but_not_done_with_progress"
+        self.request.save()
+        body = json.dumps(
+            {
+                "operations": [
+                    {"operation": "replace", "path": "/output", "value": "output"}
+                ]
+            }
+        )
+
+        response = self.fetch(
+            "/api/v1/requests/" + str(self.request.id),
+            method="PATCH",
+            body=body,
+            headers={"content-type": "application/json"},
+        )
+        self.assertEqual(200, response.code)
+        self.request.reload()
+        self.assertEqual("output", self.request.output)
+
+    def test_patch_replace_error_class(self):
+        self.request.error_class = "Klazz1"
+        body = json.dumps(
+            {
+                "operations": [
+                    {"operation": "replace", "path": "/error_class", "value": "error"}
+                ]
+            }
+        )
+        self.request.save()
+
+        response = self.fetch(
+            "/api/v1/requests/" + str(self.request.id),
+            method="PATCH",
+            body=body,
+            headers={"content-type": "application/json"},
+        )
+        self.request.reload()
+        self.assertEqual(200, response.code)
+        self.assertEqual("error", self.request.error_class)
+
+    def test_patch_replace_bad_status(self):
+        self.request.save()
+        body = json.dumps(
+            {
+                "operations": [
+                    {"operation": "replace", "path": "/status", "value": "bad"}
+                ]
+            }
+        )
+        response = self.fetch(
+            "/api/v1/requests/" + str(self.request.id),
+            method="PATCH",
+            body=body,
+            headers={"content-type": "application/json"},
+        )
+        self.assertGreaterEqual(response.code, 400)
+
+    def test_patch_update_output_for_complete_request(self):
+        self.request.status = "SUCCESS"
+        self.request.output = "old_value"
+        self.request.save()
+        body = json.dumps(
+            {
+                "operations": [
+                    {
+                        "operation": "replace",
+                        "path": "/output",
+                        "value": "shouldnt work",
+                    }
+                ]
+            }
+        )
+        response = self.fetch(
+            "/api/v1/requests/" + str(self.request.id),
+            method="PATCH",
+            body=body,
+            headers={"content-type": "application/json"},
+        )
+        self.request.reload()
+        self.assertGreaterEqual(response.code, 400)
+        self.assertEqual(self.request.output, "old_value")
+
+    def test_patch_no_system(self):
+        good_id_does_not_exist = "".join("1" for _ in range(24))
+        response = self.fetch(
+            "/api/v1/requests/" + good_id_does_not_exist,
+            method="PATCH",
+            body='{"operations": [{"operation": "fake"}]}',
+            headers={"content-type": "application/json"},
+        )
+        self.assertEqual(response.code, 404)
+
+    def test_patch_replace_bad_path(self):
+        self.request.save()
+        body = json.dumps(
+            {"operations": [{"operation": "replace", "path": "/bad", "value": "error"}]}
+        )
+        response = self.fetch(
+            "/api/v1/requests/" + str(self.request.id),
+            method="PATCH",
+            body=body,
+            headers={"content-type": "application/json"},
+        )
+        self.assertGreaterEqual(response.code, 400)
+
+    def test_patch_bad_operation(self):
+        self.request.save()
+        response = self.fetch(
+            "/api/v1/requests/" + str(self.request.id),
+            method="PATCH",
+            body='{"operations": [{"operation": "fake"}]}',
+            headers={"content-type": "application/json"},
+        )
+        self.assertGreaterEqual(response.code, 400)
+
+    def test_prometheus_endpoint(self):
+        handler = self.app.find_handler(request=Mock(path="/api/v1/requests"))
+        c = handler.handler_class(
+            self.app, Mock(path="/api/v1/requests/111111111111111111111111")
+        )
+        assert c.prometheus_endpoint == "/api/v1/requests/<ID>"
+
+    def test_update_job_numbers(self):
+        self.job.save()
+        self.request.metadata["_bg_job_id"] = str(self.job.id)
+        self.request.save()
+        body = json.dumps(
+            {
+                "operations": [
+                    {"operation": "replace", "path": "/status", "value": "SUCCESS"}
+                ]
+            }
+        )
+        response = self.fetch(
+            "/api/v1/requests/" + str(self.request.id),
+            method="PATCH",
+            body=body,
+            headers={"content-type": "application/json"},
+        )
+        self.assertEqual(response.code, 200)
+        self.job.reload()
+        self.assertEqual(self.job.success_count, 1)
+        self.assertEqual(self.job.error_count, 0)
+
+    def test_update_job_numbers_error(self):
+        self.job.save()
+        self.request.metadata["_bg_job_id"] = str(self.job.id)
+        self.request.save()
+        body = json.dumps(
+            {
+                "operations": [
+                    {"operation": "replace", "path": "/status", "value": "ERROR"}
+                ]
+            }
+        )
+        response = self.fetch(
+            "/api/v1/requests/" + str(self.request.id),
+            method="PATCH",
+            body=body,
+            headers={"content-type": "application/json"},
+        )
+        self.assertEqual(response.code, 200)
+        self.job.reload()
+        self.assertEqual(self.job.success_count, 0)
+        self.assertEqual(self.job.error_count, 1)
+
+    def test_update_job_invalid_id(self):
+        self.request.metadata["_bg_job_id"] = "".join(["1" for _ in range(24)])
+        self.request.save()
+        body = json.dumps(
+            {
+                "operations": [
+                    {"operation": "replace", "path": "/status", "value": "ERROR"}
+                ]
+            }
+        )
+        response = self.fetch(
+            "/api/v1/requests/" + str(self.request.id),
+            method="PATCH",
+            body=body,
+            headers={"content-type": "application/json"},
+        )
+        self.assertEqual(response.code, 200)
 
 
 class TestBlocking(object):
@@ -193,13 +502,13 @@ class RequestListAPITest(TestHandlerBase):
         self.request_mock.__getitem__.return_value = self.request_mock
         self.request_mock.__len__.return_value = 1
 
-        mongo_patcher = patch("brew_view.controllers.request_list_api.Request.objects")
+        mongo_patcher = patch("brew_view.handlers.v1.request.Request.objects")
         self.addCleanup(mongo_patcher.stop)
         self.mongo_mock = mongo_patcher.start()
         self.mongo_mock.count.return_value = 1
 
         serialize_patcher = patch(
-            "brew_view.controllers.request_list_api.MongoParser.serialize_request"
+            "brew_view.handlers.v1.request.MongoParser.serialize_request"
         )
         self.addCleanup(serialize_patcher.stop)
         self.serialize_mock = serialize_patcher.start()
@@ -214,7 +523,7 @@ class RequestListAPITest(TestHandlerBase):
 
         super(RequestListAPITest, self).setUp()
 
-    @patch("brew_view.controllers.request_list_api.RequestListAPI._get_query_set")
+    @patch("brew_view.handlers.v1.request.RequestListAPI._get_query_set")
     def test_get(self, get_query_set_mock):
         query_set = MagicMock()
         query_set.count.return_value = 1
@@ -232,9 +541,9 @@ class RequestListAPITest(TestHandlerBase):
         self.assertEqual("1", response.headers["recordsTotal"])
         self.assertEqual("1", response.headers["draw"])
 
-    @patch("brew_view.controllers.request_list_api.System.objects")
-    @patch("brew_view.controllers.request_list_api.MongoParser.parse_request")
-    @patch("brew_view.controllers.request_list_api.thrift_context")
+    @patch("brew_view.handlers.v1.request.System.objects")
+    @patch("brew_view.handlers.v1.request.MongoParser.parse_request")
+    @patch("brew_view.handlers.v1.request.thrift_context")
     def test_post_json(self, context_mock, parse_mock, system_mock):
         context_mock.return_value = self.fake_context
         self.client_mock.processRequest.return_value = self.future_mock
@@ -257,8 +566,8 @@ class RequestListAPITest(TestHandlerBase):
         self.assertTrue(self.request_mock.save.called)
         self.client_mock.processRequest.assert_called_once_with(self.request_mock.id)
 
-    @patch("brew_view.controllers.request_list_api.MongoParser.parse_request")
-    @patch("brew_view.controllers.request_list_api.thrift_context")
+    @patch("brew_view.handlers.v1.request.MongoParser.parse_request")
+    @patch("brew_view.handlers.v1.request.thrift_context")
     def test_post_invalid(self, context_mock, parse_mock):
         context_mock.return_value = self.fake_context
         self.client_mock.processRequest.return_value = self.future_mock
@@ -275,8 +584,8 @@ class RequestListAPITest(TestHandlerBase):
         self.assertTrue(self.request_mock.delete.called)
         self.assertTrue(self.client_mock.processRequest.called)
 
-    @patch("brew_view.controllers.request_list_api.MongoParser.parse_request")
-    @patch("brew_view.controllers.request_list_api.thrift_context")
+    @patch("brew_view.handlers.v1.request.MongoParser.parse_request")
+    @patch("brew_view.handlers.v1.request.thrift_context")
     def test_post_publishing_exception(self, context_mock, parse_mock):
         context_mock.return_value = self.fake_context
         self.client_mock.processRequest.return_value = self.future_mock
@@ -292,8 +601,8 @@ class RequestListAPITest(TestHandlerBase):
         self.assertEqual(response.code, 502)
         self.assertTrue(self.request_mock.delete.called)
 
-    @patch("brew_view.controllers.request_list_api.MongoParser.parse_request")
-    @patch("brew_view.controllers.request_list_api.thrift_context")
+    @patch("brew_view.handlers.v1.request.MongoParser.parse_request")
+    @patch("brew_view.handlers.v1.request.thrift_context")
     def test_post_exception(self, context_mock, parse_mock):
         context_mock.return_value = self.fake_context
         self.future_mock.set_exception(Exception())
@@ -317,8 +626,8 @@ class RequestListAPITest(TestHandlerBase):
         )
         self.assertEqual(response.code, 400)
 
-    @patch("brew_view.controllers.request_list_api.MongoParser.parse_request")
-    @patch("brew_view.controllers.request_list_api.thrift_context")
+    @patch("brew_view.handlers.v1.request.MongoParser.parse_request")
+    @patch("brew_view.handlers.v1.request.thrift_context")
     def test_post_instance_status_exception(self, context_mock, parse_mock):
         context_mock.return_value = self.fake_context
         self.client_mock.processRequest.return_value = self.future_mock
