+++ test/unit/handlers/v1/system_test.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 import copy
 import json
 
@@ -12,7 +13,7 @@ from bg_utils.mongo.models import System
 from brewtils.models import PatchOperation, Parameter, Choices
 from brewtils.schema_parser import SchemaParser
 from brewtils.test.comparable import assert_command_equal
-from . import TestHandlerBase
+from .. import TestHandlerBase
 
 
 @pytest.fixture(autouse=True)
@@ -151,7 +152,7 @@ class SystemAPITest(TestHandlerBase):
 
         super(SystemAPITest, self).setUp()
 
-    @patch("brew_view.controllers.system_api.MongoParser.serialize_system")
+    @patch("brew_view.handlers.v1.system.MongoParser.serialize_system")
     def test_get_system(self, serialize_mock):
         serialize_mock.return_value = "serialized_system"
 
@@ -159,7 +160,7 @@ class SystemAPITest(TestHandlerBase):
         self.assertEqual(200, response.code)
         self.assertEqual("serialized_system", response.body.decode("utf-8"))
 
-    @patch("brew_view.controllers.system_api.thrift_context")
+    @patch("brew_view.handlers.v1.system.thrift_context")
     def test_delete_system(self, context_mock):
         context_mock.return_value = self.fake_context
         self.client_mock.removeSystem.return_value = self.future_mock
@@ -169,7 +170,7 @@ class SystemAPITest(TestHandlerBase):
         self.assertEqual(204, response.code)
         self.client_mock.removeSystem.assert_called_once_with("id")
 
-    @patch("brew_view.controllers.system_api.thrift_context")
+    @patch("brew_view.handlers.v1.system.thrift_context")
     def test_delete_system_thrift_exception(self, context_mock):
         context_mock.return_value = self.fake_context
         self.client_mock.removeSystem.return_value = self.future_mock
@@ -179,8 +180,8 @@ class SystemAPITest(TestHandlerBase):
         self.assertNotEqual(204, response.code)
         self.client_mock.removeSystem.assert_called_once_with("id")
 
-    @patch("brew_view.controllers.system_api.MongoParser.parse_command", Mock())
-    @patch("brew_view.controllers.system_api.MongoParser.serialize_system")
+    @patch("brew_view.handlers.v1.system.MongoParser.parse_command", Mock())
+    @patch("brew_view.handlers.v1.system.MongoParser.serialize_system")
     def test_patch_replace_commands_ok(self, serialize_mock):
         body = json.dumps(
             {
@@ -201,8 +202,8 @@ class SystemAPITest(TestHandlerBase):
         self.assertEqual("serialized_system", response.body.decode("utf-8"))
         self.assertTrue(self.system_mock.upsert_commands.called)
 
-    @patch("brew_view.controllers.system_api.MongoParser.parse_command", Mock())
-    @patch("brew_view.controllers.system_api.MongoParser.serialize_system")
+    @patch("brew_view.handlers.v1.system.MongoParser.parse_command", Mock())
+    @patch("brew_view.handlers.v1.system.MongoParser.serialize_system")
     def test_patch_replace_commands_bad(self, serialize_mock):
         self.system_mock.commands = ["a command"]
         body = json.dumps(
@@ -223,8 +224,8 @@ class SystemAPITest(TestHandlerBase):
         self.assertEqual(400, response.code)
         self.assertFalse(self.system_mock.upsert_commands.called)
 
-    @patch("brew_view.controllers.system_api.thrift_context")
-    @patch("brew_view.controllers.system_api.MongoParser.serialize_system")
+    @patch("brew_view.handlers.v1.system.thrift_context")
+    @patch("brew_view.handlers.v1.system.MongoParser.serialize_system")
     def test_patch_reload(self, serialize_mock, context_mock):
         context_mock.return_value = self.fake_context
         self.client_mock.reloadSystem.return_value = self.future_mock
@@ -241,8 +242,8 @@ class SystemAPITest(TestHandlerBase):
         self.assertEqual("serialized_system", response.body.decode("utf-8"))
         self.client_mock.reloadSystem.assert_called_once_with("id")
 
-    @patch("brew_view.controllers.system_api.MongoParser.parse_command", Mock())
-    @patch("brew_view.controllers.system_api.MongoParser.serialize_system")
+    @patch("brew_view.handlers.v1.system.MongoParser.parse_command", Mock())
+    @patch("brew_view.handlers.v1.system.MongoParser.serialize_system")
     def test_patch_update_metadata(self, serialize_mock):
         self.system_mock.metadata = {"foo": "baz"}
         body = json.dumps(
@@ -268,8 +269,8 @@ class SystemAPITest(TestHandlerBase):
         self.assertEqual("serialized_system", response.body.decode("utf-8"))
         self.assertEqual(self.system_mock.metadata, {"foo": "bar"})
 
-    @patch("brew_view.controllers.system_api.MongoParser.parse_command", Mock())
-    @patch("brew_view.controllers.system_api.MongoParser.serialize_system")
+    @patch("brew_view.handlers.v1.system.MongoParser.parse_command", Mock())
+    @patch("brew_view.handlers.v1.system.MongoParser.serialize_system")
     def test_patch_replace_description(self, serialize_mock):
         self.system_mock.description = "old_description"
         body = json.dumps(
@@ -295,8 +296,8 @@ class SystemAPITest(TestHandlerBase):
         self.assertEqual("serialized_system", response.body.decode("utf-8"))
         self.assertEqual(self.system_mock.description, "new_description")
 
-    @patch("brew_view.controllers.system_api.MongoParser.parse_command", Mock())
-    @patch("brew_view.controllers.system_api.MongoParser.serialize_system")
+    @patch("brew_view.handlers.v1.system.MongoParser.parse_command", Mock())
+    @patch("brew_view.handlers.v1.system.MongoParser.serialize_system")
     def test_patch_replace_null_empty_string(self, serialize_mock):
         self.system_mock.description = "old_description"
         body = json.dumps(
@@ -318,8 +319,8 @@ class SystemAPITest(TestHandlerBase):
         self.assertEqual("serialized_system", response.body.decode("utf-8"))
         self.assertEqual(self.system_mock.description, "")
 
-    @patch("brew_view.controllers.system_api.MongoParser.parse_command", Mock())
-    @patch("brew_view.controllers.system_api.MongoParser.serialize_system")
+    @patch("brew_view.handlers.v1.system.MongoParser.parse_command", Mock())
+    @patch("brew_view.handlers.v1.system.MongoParser.serialize_system")
     def test_patch_invalid_path_for_update(self, serialize_mock):
         body = json.dumps(
             {
@@ -373,3 +374,78 @@ class SystemAPITest(TestHandlerBase):
             headers={"content-type": "application/json"},
         )
         self.assertGreaterEqual(response.code, 400)
+
+
+class SystemListAPITest(TestHandlerBase):
+    def setUp(self):
+        self.system_mock = Mock(name="System Mock")
+        self.system_mock.filter.return_value = self.system_mock
+        self.system_mock.order_by.return_value = self.system_mock
+
+        mongo_patcher = patch("mongoengine.queryset.manager.QuerySetManager.__get__")
+        self.addCleanup(mongo_patcher.stop)
+        self.get_mock = mongo_patcher.start()
+        self.get_mock.return_value = self.system_mock
+
+        serialize_patcher = patch(
+            "brew_view.handlers.v1.system.MongoParser.serialize_system"
+        )
+        self.addCleanup(serialize_patcher.stop)
+        self.serialize_mock = serialize_patcher.start()
+        self.serialize_mock.return_value = "serialized_system"
+
+        super(SystemListAPITest, self).setUp()
+
+    def test_get(self):
+        response = self.fetch("/api/v1/systems")
+        self.assertEqual(200, response.code)
+        self.assertEqual(
+            self.serialize_mock.return_value, response.body.decode("utf-8")
+        )
+        self.assertIn(self.system_mock, self.serialize_mock.call_args[0])
+
+    def test_get_exclude_commands(self):
+        response = self.fetch("/api/v1/systems?include_commands=False")
+        self.assertEqual(200, response.code)
+        self.assertEqual(
+            self.serialize_mock.return_value, response.body.decode("utf-8")
+        )
+        self.assertEqual(self.serialize_mock.call_args[1]["exclude"], {"commands"})
+
+    def test_get_with_filter_param(self):
+        self.fetch("/api/v1/systems?name=bar")
+        self.system_mock.filter.assert_called_once_with(name="bar")
+
+    def test_get_with_filter_params(self):
+        self.fetch("/api/v1/systems?name=bar&version=1.0.0")
+        self.system_mock.filter.assert_called_once_with(name="bar", version="1.0.0")
+
+    def test_get_ignore_bad_filter_params(self):
+        self.fetch("/api/v1/systems?foo=bar")
+        self.system_mock.filter.assert_called_once_with()
+
+    @patch("bg_utils.mongo.models.System.find_unique", Mock(return_value=False))
+    @patch("brew_view.handlers.v1.system.SystemListAPI._create_new_system")
+    @patch("brew_view.handlers.v1.system.MongoParser.parse_system")
+    def test_post_new_system(self, parse_mock, create_mock):
+        parse_mock.return_value = self.system_mock
+        create_mock.return_value = Mock(), 201
+
+        response = self.fetch("/api/v1/systems", method="POST", body="")
+        self.assertEqual(201, response.code)
+        create_mock.assert_called_once_with(self.system_mock)
+
+    @patch("bg_utils.mongo.models.System.find_unique")
+    @patch(
+        "brew_view.handlers.v1.system.SystemListAPI._update_existing_system"
+    )
+    @patch("brew_view.handlers.v1.system.MongoParser.parse_system")
+    def test_post_existing_system(self, parse_mock, update_mock, find_mock):
+        parse_mock.return_value = self.system_mock
+        db_system_mock = Mock()
+        find_mock.return_value = db_system_mock
+        update_mock.return_value = Mock(), 200
+
+        response = self.fetch("/api/v1/systems", method="POST", body="")
+        self.assertEqual(200, response.code)
+        update_mock.assert_called_once_with(db_system_mock, self.system_mock)
